Consolidate duplicate mock_docker_backend fixture
- Moved shared mock_docker_backend fixture to tests/conftest.py
- Removed duplicate fixtures from test_advanced_security_scenarios.py and test_security_scanner.py
- Applied fixture only to security test modules using pytestmark to avoid interfering with other tests
- Fixed issue where autouse=True was breaking language handler factory tests
- All tests now passing: 30 language handler tests + security tests working correctly
- Eliminates code duplication and improves maintainability
